Move massage result functions from sessions_v2 to outcomes (#108598)
massage_sessions_result and massage_sessions_result_summary are only
called by outcomes code in production. The summary function is entirely
outcomes-specific (hardcodes outcome/category groupby keys). Move both
functions along with their helpers (_split_rows_groupby,
dropped_outcomes, _CategoryStats, _Project, _Period) to
snuba/outcomes.py where they belong.
